Submission

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        double result = toBeDivided / 5.0;
+        int result = toBeDivided / 5;
         System.out.println(result);
     }
 }
